GhostFactory generates ghosts. Ghost colisions only stop the player a…

+++ project/app/src/main/java/edu/ub/pis2019/pis_16/tempac/Model/Factory.kt
@@ -1,5 +1,5 @@
 package edu.ub.pis2019.pis_16.tempac.Model
 
 interface Factory {
-    fun create(temperature: Int): Ghost
+    fun create(temperature: Float): Ghost
 }
